# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "recsys",
    "description": "Precomputed e-commerce recommendations (Retailrocket). See docs/API.md.",
    "version": "1.0"
  },
  "paths": {
    "/recommendations/{user_id}": {
      "get": {
        "summary": "Recommendations",
        "operationId": "recommendations_recommendations__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "User Id"
            }
          },
          {
            "name": "n",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "N"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/similar/{item_id}": {
      "get": {
        "summary": "Similar",
        "operationId": "similar_similar__item_id__get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "Item Id"
            }
          },
          {
            "name": "n",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "N"
            }
          },
          {
            "name": "available_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Available Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimilarItemsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Health": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "redis": {
            "type": "boolean",
            "title": "Redis"
          },
          "qdrant": {
            "type": "boolean",
            "title": "Qdrant"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "snapshot": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot"
          }
        },
        "type": "object",
        "required": [
          "status",
          "redis",
          "qdrant",
          "version"
        ],
        "title": "Health"
      },
      "RecommendationsResponse": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ScoredItem"
            },
            "type": "array",
            "title": "Items"
          },
          "source": {
            "$ref": "#/components/schemas/Source"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version",
            "description": "Snapshot version served (null when degraded)"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "items",
          "source",
          "version"
        ],
        "title": "RecommendationsResponse"
      },
      "ScoredItem": {
        "properties": {
          "item_id": {
            "type": "integer",
            "title": "Item Id"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "rank": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Rank"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "score",
          "rank"
        ],
        "title": "ScoredItem"
      },
      "SimilarItemsResponse": {
        "properties": {
          "item_id": {
            "type": "integer",
            "title": "Item Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ScoredItem"
            },
            "type": "array",
            "title": "Items"
          },
          "collection": {
            "type": "string",
            "title": "Collection"
          },
          "available_only": {
            "type": "boolean",
            "title": "Available Only"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "items",
          "collection",
          "available_only"
        ],
        "title": "SimilarItemsResponse"
      },
      "Source": {
        "type": "string",
        "enum": [
          "personalized",
          "popular",
          "popular_degraded"
        ],
        "title": "Source"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}